more install script fixes to handle systemd ourselves and not enable daemons on install

+++ .gopythongo/post_rm.sh
@@ -1,17 +1,21 @@
 #!/usr/bin/env bash
 
-# fpm wraps this in a shell function and ignores the "purge" action. So basically we
-# handle any invocation like a purge
-deluser --system authserver
-delgroup --system authserver
+if [ "$1" = "purge" ]; then
+    deluser --system authserver
+    delgroup --system authserver
 
-if [ -x "/usr/bin/deb-systemd-helper" ]; then
-    /usr/bin/deb-systemd-helper purge authserver.service >/dev/null
-    /usr/bin/deb-systemd-helper unmask authserver.service >/dev/null
-    /usr/bin/deb-systemd-helper purge dkimsigner.service >/dev/null
-    /usr/bin/deb-systemd-helper unmask dkimsigner.service >/dev/null
-fi
+    if [ -x "/usr/bin/deb-systemd-helper" ]; then
+        /usr/bin/deb-systemd-helper purge authserver.service >/dev/null
+        /usr/bin/deb-systemd-helper unmask authserver.service >/dev/null
+        /usr/bin/deb-systemd-helper purge dkimsigner.service >/dev/null
+        /usr/bin/deb-systemd-helper unmask dkimsigner.service >/dev/null
+    fi
 
-if [ -x /etc/appconfig/authserver ]; then
-    rm -rf /etc/appconfig/authserver
+    if [ -x /etc/appconfig/authserver ]; then
+        rm -rf /etc/appconfig/authserver
+    fi
+
+    if [ -x /etc/appconfig/dkimsigner ]; then
+        rm -rf /etc/appconfig/dkimsigner
+    fi
 fi
